drm: Handle u32 wrapping in nouveau_fence_done
The naive comparison

  if (fctx->read(chan) < fence->sequence)

to determine whether a fence is expired or not will fail when the
sequence number wraps around. Fix this by using a wrapping-safe
comparison.

Also wrap fctx->read as nouveau_fence_current in the nouveau_fence
interface to allow other nouveau components to monitor progress
(intended to be used in debug prints).

Signed-off-by: Lauri Peltonen <lpeltonen@nvidia.com>